Cleanup for zmq
